pyproject.toml
[project]
name = "pyawn"
version = "0.1.0"
description = "CLI para treinar aberturas de xadrez baseadas em datasets históricos e Stockfish"
authors = [
    {name = "Vitor Meriat",email = "vitormeriat@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "click>=8.1",
    "python-chess>=1.10",
    "rich>=13.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-cov>=5.0",
    "ruff>=0.4",
    "mypy>=1.10",
]

[project.scripts]
pyawn = "pyawn.cli:cli"

[tool.poetry]
packages = [{include = "pyawn", from = "src"}]

[tool.pytest.ini_options]
testpaths = ["tests"]

[tool.ruff]
target-version = "py312"
line-length = 100

[tool.mypy]
python_version = "3.12"
strict = true

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
